+++ src/common/bip85/common_bip85.h
@@ -49,8 +49,6 @@ void bolos_ux_bip85_hex(uint8_t* hex_out, uint8_t num_bytes, unsigned int index)
  * @param[out] pwd     Pointer to a buffer where the generated BIP85 Base64 password will be stored.
  * @param[in]  pwd_len Length of the password in bytes.
  * @param[in]  index   Index to be used in the BIP32 path.
- *
- * @return The number of bytes written to the output buffer.
  */
 uint8_t bolos_ux_bip85_pwd_base64(char* pwd, uint8_t pwd_len, unsigned int index);
 
@@ -65,8 +63,10 @@ uint8_t bolos_ux_bip85_pwd_base64(char* pwd, uint8_t pwd_len, unsigned int index
  * @param[out] pwd     Pointer to a buffer where the generated BIP85 Base85 password will be stored.
  * @param[in]  pwd_len Length of the password in bytes.
  * @param[in]  index   Index to be used in the BIP32 path.
+ *
+ * @return The number of bytes written to the output buffer.
  */
-void bolos_ux_bip85_pwd_base85(char* pwd, uint8_t pwd_len, unsigned int index);
+uint8_t bolos_ux_bip85_pwd_base85(char* pwd, uint8_t pwd_len, unsigned int index);
 
 /**
  * @brief Generates a series of random dice rolls using BIP85.
@@ -78,5 +78,7 @@ void bolos_ux_bip85_pwd_base85(char* pwd, uint8_t pwd_len, unsigned int index);
  * @param[in]  sides  Number of sides on each die (must be between 2 and UINT32_MAX >> 1).
  * @param[in]  rolls  Number of dice rolls to generate (must be between 1 and UINT32_MAX >> 1).
  * @param[in]  index  Index to be used in the BIP32 path.
+ *
+ * @return The number of bytes written to the output buffer.
  */
 void bolos_ux_bip85_dice(uint32_t* out, uint32_t sides, uint32_t rolls, unsigned int index);