+++ src/constants.h
@@ -20,9 +20,9 @@
 #define ONBOARDING_WORD_COMPLETION_MAX_ITEMS 8
 #define BOLOS_UX_HASH_LENGTH                 4  // as on the blue
 
-#define BIP39_MAX_WORD_LENGTH          8
-#define SSKR_MAX_NUMBER_LENGTH         2
-#define BIP85_INDEX_MAX_NUMBER_LENGTH  7
+#define BIP39_MAX_WORD_LENGTH         8
+#define SSKR_MAX_NUMBER_LENGTH        2
+#define BIP85_INDEX_MAX_NUMBER_LENGTH 7
 
 enum __attribute__((packed)) {
     BIP39_MNEMONIC_SIZE_12 = 12,