+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add unit tests for base64 and base85
 - Add BIP85 application menus to generate BIP39 children on Flex and Stax
 - Add BIP85 application menus to generate base64 passwords on Flex and Stax
+- Add BIP85 application menus to generate base85 passwords on Flex and Stax
 
 ## [1.8.1] - 2025-07-24